I added more OnlMonClient classes (OnlMonHodo, OnlMonProp, OnlMonTrigSig and OnlMonV1495).  As I checked the outputs of these new classes with run 27400, I modified some other files (mostly to fix bugs).
The SQHit class was updated to hold the level info.

The default geometry schema in JobOptsSvc.cxx was updated so that D0 was taken into the mapping.

The channel-mapper classes were updated so as to not miss (or mis-map) some detectors.  Particularly the detector name for the X-T curve of the prop tube is now correctly handled.