An interim version of the real-time event display.
I copied several macros from "e1039-analysis/CODAChainDev" to "packages/Display/macro" and modified them.  It now has a function (button) of moving back to previous events.  This function uses in backend another new function of re-opening DST files.  So it is nearly ready for auto-switching to the next run in the real time mode.

I made a few changes in "Fun4AllDstOutputManager" and "PHNodeIOManager" and created "EvtDispFilter".  By using them, "Fun4MainDaq.C" can output a small DST file for the event display.  As the output is flushed every event (via TTree::AutoSave()), it is readable when being written.