Temporarily removed the PatternDB work
Since we do not have man power to continue development or maintenance of this code, it is temporarily moved to a separate directory to reduce the compile time. It is still kept in the repo so that people can pick it up again in the near future. The track finder inside KalmanFastTrackingWrapper is changed back to KalmanFastTracking